sentriscope-mvp/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36,<2.1
psycopg[binary]==3.2.3
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.9,<3
pydantic-settings==2.1.0
python-dotenv==1.0.0
